Add proposal analytics: total votes, participation rate, and voting trends for proposals
