+++ icn-blockchain/src/governance.rs
@@ -1,4 +1,4 @@
-use std::collections::HashMap;
+use std::collections::{HashMap, VecDeque};  // Added VecDeque here
 use chrono::{Utc, Duration};
 
 #[derive(Debug, Clone, PartialEq)]
@@ -75,6 +75,73 @@ impl Proposal {
 
         if self.status == ProposalStatus::Open && time_left <= reminder_threshold {
             println!("Reminder: Proposal '{}' is closing soon! Time left: {} seconds", self.description, time_left);
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct ProposalHistory {
+    proposals: VecDeque<Proposal>,
+}
+
+impl ProposalHistory {
+    pub fn new() -> Self {
+        ProposalHistory {
+            proposals: VecDeque::new(),
+        }
+    }
+
+    pub fn add_proposal(&mut self, proposal: Proposal) {
+        self.proposals.push_back(proposal);
+    }
+
+    pub fn display_history(&self) {
+        for proposal in &self.proposals {
+            println!(
+                "Proposal ID: {}, Description: '{}', Result: {}, Total Votes: {}",
+                proposal.id,
+                proposal.description,
+                match proposal.status {
+                    ProposalStatus::Closed => "Closed",
+                    ProposalStatus::Open => "Open",
+                },
+                proposal.total_votes()
+            );
+        }
+    }
+
+    /// Calculates the total votes cast on each proposal
+    pub fn total_votes_for_proposals(&self) {
+        for proposal in &self.proposals {
+            let total_votes = proposal.total_votes();
+            println!(
+                "Proposal ID: {} - Total Votes Cast: {}",
+                proposal.id, total_votes
+            );
+        }
+    }
+
+    /// Calculates participation rate based on a provided total member count
+    pub fn participation_rate(&self, total_members: usize) {
+        for proposal in &self.proposals {
+            let unique_voters = proposal.votes.len();
+            let participation_rate = (unique_voters as f64 / total_members as f64) * 100.0;
+            println!(
+                "Proposal ID: {} - Participation Rate: {:.2}%",
+                proposal.id, participation_rate
+            );
+        }
+    }
+
+    /// Calculates positive vs. negative vote trend for each proposal
+    pub fn voting_trends(&self) {
+        for proposal in &self.proposals {
+            let positive_votes: i64 = proposal.votes.values().filter(|&&v| v > 0).sum();
+            let negative_votes: i64 = proposal.votes.values().filter(|&&v| v < 0).sum();
+            println!(
+                "Proposal ID: {} - Positive Votes: {}, Negative Votes: {}",
+                proposal.id, positive_votes, negative_votes
+            );
         }
     }
 }
